+++ src/ios/OSInAppBrowserError.swift
@@ -1,11 +1,13 @@
 enum OSInAppBrowserTarget {
-    case openInExternalBrowser
-    case openInSystemBrowser
+    case externalBrowser
+    case systemBrowser
+    case webView
 }
 
 enum OSInAppBrowserError: Error {
     case inputArgumentsIssue(target: OSInAppBrowserTarget)
     case failedToOpen(url: String, onTarget: OSInAppBrowserTarget)
+    case noBrowserToClose
     
     private var code: Int {
         let result: Int
@@ -13,6 +15,7 @@ enum OSInAppBrowserError: Error {
         switch self {
         case .inputArgumentsIssue: result = 0
         case .failedToOpen: result = 0
+        case .noBrowserToClose: result = 0
         }
         
         return result
@@ -26,8 +29,9 @@ enum OSInAppBrowserError: Error {
             let targetString: String
             
             switch target {
-            case .openInExternalBrowser: targetString = "openInExternalBrowser"
-            case .openInSystemBrowser: targetString = "openInSystemBrowser"
+            case .externalBrowser: targetString = "openInExternalBrowser"
+            case .systemBrowser: targetString = "openInSystemBrowser"
+            case .webView: targetString = "openInWebView"
             }
             
             result = "The input parameters for '\(targetString)' are invalid."
@@ -35,11 +39,14 @@ enum OSInAppBrowserError: Error {
             let targetString: String
             
             switch target {
-            case .openInExternalBrowser: targetString = "Safari"
-            case .openInSystemBrowser: targetString = "SFSafariViewController"
+            case .externalBrowser: targetString = "Safari"
+            case .systemBrowser: targetString = "SFSafariViewController"
+            case .webView: targetString = "WebView"
             }
             
             result = "Couldn't open '\(url)' using \(targetString)."
+        case .noBrowserToClose:
+            result = "No browser view to close."
         }
         
         return result