+++ src/www/definitions.ts
@@ -59,11 +59,10 @@ export interface iOSWebViewOptions {
 
   enableViewportScale: boolean;
   allowInLineMediaPlayback: boolean;
-  keyboardDisplayRequiresUserAction: boolean;
   surpressIncrementalRendering: boolean;
 
   viewStyle: iOSViewStyle;
-  animation: iOSAnimation;
+  animationEffect: iOSAnimation;
 }
 
 export interface AndroidWebViewOptions {
@@ -115,7 +114,7 @@ export interface BrowserCallbacks {
 }
 
 export enum CallbackEvent {
-  SUCCESS,
-  PAGE_CLOSED,
-  PAGE_LOAD_COMPLETED
+  SUCCESS = 1,
+  PAGE_CLOSED = 2,
+  PAGE_LOAD_COMPLETED = 3
 }